Fix faulty logic in subtraction operations
The carry bit is set when the unsigned value of the value being
subtracted is greater than the unsigned value of the accumulator.

The carry flag was not being set when the value being subtracted
was 0. This commit also simplifies how the Carry, Zero and Aux
Carry flags are set in each subtraction function.